wasm-bot/requirements.txt
streamlit
pandasai
pandasai[excel]
python-dotenv
watchdog
cchardet
seaborn
matplotlib
numpy
statsmodels
pydeck
plotly.express
folium
streamlit_folium
streamlit-elements
streamlit-tags
streamlit-aggrid
streamlit-option-menu
pillow
pandasai
openai
bokeh
